pyproject.toml
[tool.poetry]
name = "LOL_counter"
version = "0.1.0"
description = ""
authors = ["every"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

# --- LangChain 및 LLM 관련 라이브러리 ---
langchain-openai = "^1.0.2"      # OpenAI 모델(GPT 등)을 LangChain과 연동
langchain-core = "^1.0.4"        # LangChain의 핵심 추상화 및 구성 요소 (프롬프트, 파서 등)
python-dotenv = "^1.2.1"         # .env 파일에서 환경 변수를 로드
pydantic = "^2.12.4"             # 데이터 모델 정의 및 검증 (LangServe 입력/출력 스키마 정의에 사용)
requests = "^2.32.5"
beautifulsoup4 = "^4.14.2"
streamlit = "^1.33.0"
streamlit-chat = "^0.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
